flask_app/requirements.txt
Flask
pymongo
flask_swagger_ui
python-dotenv